refactor(loggerService): Changed class for function methods.
